+++ multigtfs/tests/shape.py
@@ -15,9 +15,10 @@
 
 import StringIO
 
+from django.contrib.gis.geos import MultiLineString
 from django.test import TestCase
 
-from multigtfs.models import Feed, Shape, ShapePoint
+from multigtfs.models import Feed, Route, Shape, ShapePoint, Trip
 
 
 class ShapeTest(TestCase):
@@ -63,10 +64,10 @@ S1,36.425288,-117.133162,1
         shape = Shape.objects.get()
         self.assertEqual(shape.feed, self.feed)
         self.assertEqual(shape.shape_id, 'S1')
+        self.assertEqual(shape.geometry, None)
         shape_pt = ShapePoint.objects.get()
         self.assertEqual(shape_pt.shape, shape)
-        self.assertEqual(str(shape_pt.lat), '36.425288')
-        self.assertEqual(str(shape_pt.lon), '-117.133162')
+        self.assertEqual(shape_pt.point.coords, (-117.133162, 36.425288))
         self.assertEqual(shape_pt.sequence, 1)
         self.assertEqual(shape_pt.traveled, None)
 
@@ -79,10 +80,10 @@ S1,36.425288,-117.133162,1,0
         shape = Shape.objects.get()
         self.assertEqual(shape.feed, self.feed)
         self.assertEqual(shape.shape_id, 'S1')
+        self.assertEqual(shape.geometry, None)
         shape_pt = ShapePoint.objects.get()
         self.assertEqual(shape_pt.shape, shape)
-        self.assertEqual(str(shape_pt.lat), '36.425288')
-        self.assertEqual(str(shape_pt.lon), '-117.133162')
+        self.assertEqual(shape_pt.point.coords, (-117.133162, 36.425288))
         self.assertEqual(shape_pt.sequence, 1)
         self.assertEqual(shape_pt.traveled, 0)
 
@@ -95,10 +96,10 @@ S1,36.425288,-117.133162,1,
         shape = Shape.objects.get()
         self.assertEqual(shape.feed, self.feed)
         self.assertEqual(shape.shape_id, 'S1')
+        self.assertEqual(shape.geometry, None)
         shape_pt = ShapePoint.objects.get()
         self.assertEqual(shape_pt.shape, shape)
-        self.assertEqual(str(shape_pt.lat), '36.425288')
-        self.assertEqual(str(shape_pt.lon), '-117.133162')
+        self.assertEqual(shape_pt.point.coords, (-117.133162, 36.425288))
         self.assertEqual(shape_pt.sequence, 1)
         self.assertEqual(shape_pt.traveled, None)
 
@@ -126,3 +127,21 @@ S1,36.425288,-117.133162,1
 shape_id,shape_pt_lat,shape_pt_lon,shape_pt_sequence,shape_dist_traveled
 S1,36.425288,-117.133162,1,1.1
 """)
+
+    def test_update_geometry_on_shapepoint_save(self):
+        shape = Shape.objects.create(feed=self.feed)
+        route = Route.objects.create(feed=self.feed, rtype=3)
+        trip = Trip.objects.create(shape=shape, route=route)
+        ShapePoint.objects.create(
+            shape=shape, point="POINT(-117.133162 36.425288)", sequence=1)
+        ShapePoint.objects.create(
+            shape=shape, point="POINT(-117.13 36.42)", sequence=2)
+
+        shape = Shape.objects.get(id=shape.id)
+        trip = Trip.objects.get(id=trip.id)
+        route = Route.objects.get(id=route.id)
+        self.assertEqual(
+            shape.geometry.coords,
+            ((-117.133162, 36.425288), (-117.13, 36.42)))
+        self.assertEqual(trip.geometry, shape.geometry)
+        self.assertEqual(route.geometry, MultiLineString(shape.geometry))