+++ CHANGELOG.rst
@@ -1,9 +1,19 @@
 Changelog
 =========
 
+0.3.1 (2014-03-12)
+------------------
+This release was generously sponsored by MRCagney.
+
+* Add example project 'explore', which represents a feed as linked pages with
+  OpenLayer maps.
+* Add cached geometry for Routes, Trips, and Shapes.
+* Extend fields for real-world data (FeedInfo.version, Route.short_name).
+* Drop support for Points as geography fields.
+
 0.3.0 (2014-02-01)
 ------------------
-This release was sponsored by MRCagney_.
+This release was generously sponsored by MRCagney.
 
 * Convert to GeoDjango: Stops and ShapePoints use Points rather than lat/long,
   admin shows map of points, and new configuration settings to customize.
@@ -31,4 +41,3 @@ This release was sponsored by MRCagney_.
 * Fixed Fare.transfers for unlimited rides (use None instead of -1)
 * First PyPi version
 
-.. _MRCagney: http://mrcagney.co.nz